==> code/notebooks/00_importing_data.ipynb <==
# NFL Betting Odds

**Author:** Jakob Malasig
**Last modified:** 12-17-2024 (created: 11-17-2024)

**Description:** I will be downloading data for NFL seasons from Kaggle and cleaning the raw data. Will then export it as a csv file for further anaylsis.

<br>

Import libraries:

import pandas as pd

<br>

Load data:

# load all of Fixture Download data into dataframes
# odds
nfl_odds = pd.read_csv("../../data/raw/spreadspoke_scores.csv", encoding = "UTF-8")

# team data
nfl_teams = pd.read_csv("../../data/raw/nfl_teams.csv", encoding = "UTF-8")

display(nfl_odds)
display(nfl_teams)

<br>

Cleaning data for nfl odds df:

# remove unwanted columns
columns_to_drop = ["Unnamed: 17", "Unnamed: 18", "Unnamed: 19", "Unnamed: 20", "Unnamed: 21", "Unnamed: 22", "Unnamed: 23", "Unnamed: 24"]
nfl_odds = nfl_odds.drop(columns = columns_to_drop)

# remove records (keep only seasons from 2018-2024)
nfl_odds = nfl_odds[nfl_odds["schedule_season"] >= 2018]

# remove records (keep only regular season games)
nfl_odds = nfl_odds[nfl_odds["schedule_playoff"] == False]

# remove records (keep only games played)
nfl_odds = nfl_odds.dropna(subset = ["score_home"])

# drop unneeded columns anymore
columns_to_drop = ["schedule_week", "schedule_playoff", "stadium_neutral"]
nfl_odds = nfl_odds.drop(columns = columns_to_drop)

# convert data types in columns
nfl_odds["score_home"] = nfl_odds["score_home"].astype(int)
nfl_odds["score_away"] = nfl_odds["score_away"].astype(int)
nfl_odds["over_under_line"] = nfl_odds["over_under_line"].astype(float)

# convert the date column
nfl_odds["schedule_date"] = pd.to_datetime(nfl_odds["schedule_date"], format = "%m/%d/%Y")
nfl_odds["schedule_date"] = nfl_odds["schedule_date"].dt.strftime("%m/%d/%Y")

nfl_odds.dtypes

# rename columns for easy of integrating
columns_to_rename = {
    "schedule_date": "date",
    "stadium": "location",
    "team_home": "home_team",
    "team_away": "away_team"
}
nfl_odds = nfl_odds.rename(columns = columns_to_rename)

display(nfl_odds)

<br>

Cleaning nfl teams dataframe:

columns_to_drop = ["team_name_short", "team_id_pfr", "team_conference_pre2002", "team_division_pre2002", "team_conference", "team_division"]
nfl_teams = nfl_teams.drop(columns = columns_to_drop)

nfl_teams = nfl_teams[nfl_teams["team_name"] != "St. Louis Cardinals"]
nfl_teams = nfl_teams[nfl_teams["team_name"] != "Phoenix Cardinals"]

nfl_teams = nfl_teams.groupby("team_id").last().reset_index()

display(nfl_teams)

<br>

Add favorite NFL team column:

# add the full team name as a favorite team column
nfl_odds = nfl_odds.merge(nfl_teams, left_on = "team_favorite_id", right_on = "team_id", how = "left")

nfl_odds = nfl_odds.rename(columns = {"team_name": "team_favorite"})
nfl_odds.drop("team_favorite_id", axis = 1)

new_column_order = ["date", 
                    "schedule_season", 
                    "home_team", 
                    "score_home", 
                    "away_team",
                    "score_away", 
                    "team_favorite",
                    "spread_favorite",
                   "over_under_line",
                   "location",
                   "weather_temperature",
                   "weather_humidity",
                   "weather_detail"]
nfl_odds = nfl_odds[new_column_order]

# change the spread to positive to help with analysis
nfl_odds["spread_favorite"] = nfl_odds["spread_favorite"].abs()

display(nfl_odds)

Creating new csv file from cleaned data:

nfl_odds.to_csv("../../data/clean/clean_kaggle_data.csv", encoding = "UTF-8", index = False)
