==> code/notebooks/02_integrating_data.ipynb <==
# NFL Betting Odds

**Author:** Jakob Malasig
**Last modified:** 12-17-2024 (created: 12-17-2024)

**Description:** I will be integrating the data I downloaded/scraped into one dataset.

Import libraries:

import pandas as pd

<br>

Import data:

# load data into dataframes
# odds
nfl_odds = pd.read_csv("../../data/clean/clean_kaggle_data.csv", encoding = "UTF-8")

# nfl data
nfl_data = pd.read_csv("../../data/clean/clean_fixture_download.csv", encoding = "UTF-8")

display(nfl_odds)
display(nfl_data)

<br>

Integrate the datasets:

full_nfl_data = nfl_odds.merge(nfl_data, how = "inner", on = ["date", "location", "home_team", "away_team"])

display(full_nfl_data)

<br>

Export as csv:

full_nfl_data.to_csv("../../data/final/nfl_data_for_analysis.csv", encoding = "UTF-8", index = False)
